requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.4.2
yfinance>=0.2.31
requests>=2.31.0
langchain-community>=0.0.13